fix(cookbook/): ajustando  execucao dos cookbooks para async tasks
